vdisk: flush disk before loading new one into drive
Flush will occur before the load dialog is presented, as it's not
unheard of for users to insert the same disk.  The logic for replacing a
disk is to try loadind the new one before deciding whether to eject the
old one - if the disk had not been saved before it was reselectd to
insert, we'd end up with a mismatch between on-disk and in-memory.

Also always mark a newly-created disk as "dirty" so that something will
be saved even if the new disk is not formatted.