pass dialog title to file requester, not list of extensions
I do want the ability to add filters, but it's not been done yet...

So for now, switch it for something else I've been wanting to do: custom
dialog titles.

The CLI file requester uses the title as the basis for a prompt.

The Mac OS X+ code is untested.